dispatcher/requirements.txt
boto3==1.24.45
botocore==1.27.45
colorama==0.4.5
jmespath==1.0.1
py-cpuinfo==8.0.0
python-dateutil==2.8.2
s3transfer==0.6.0
shortuuid==1.0.9
six==1.16.0
urllib3==1.26.11
